Remove abstract class LargeSingleFormatter
