+++ src/Single/LargeSinglePlainTextFormatter.php
@@ -2,11 +2,66 @@
 
 namespace CultuurNet\CalendarSummaryV3\Single;
 
+use CultuurNet\CalendarSummaryV3\Translator;
 use CultuurNet\SearchV3\ValueObjects\Offer;
 use DateTimeInterface;
+use IntlDateFormatter;
 
-class LargeSinglePlainTextFormatter extends LargeSingleFormatter implements SingleFormatterInterface
+class LargeSinglePlainTextFormatter implements SingleFormatterInterface
 {
+    /**
+     * @var IntlDateFormatter
+     */
+    protected $fmt;
+
+    /**
+     * @var IntlDateFormatter
+     */
+    protected $fmtWeekDayLong;
+
+    /**
+     * @var IntlDateFormatter
+     */
+    protected $fmtTime;
+
+    /**
+     * @var Translator
+     */
+    protected $trans;
+
+    public function __construct(string $langCode)
+    {
+        $this->fmt = new IntlDateFormatter(
+            $langCode,
+            IntlDateFormatter::FULL,
+            IntlDateFormatter::FULL,
+            date_default_timezone_get(),
+            IntlDateFormatter::GREGORIAN,
+            'd MMMM yyyy'
+        );
+
+        $this->fmtWeekDayLong = new IntlDateFormatter(
+            $langCode,
+            IntlDateFormatter::FULL,
+            IntlDateFormatter::FULL,
+            date_default_timezone_get(),
+            IntlDateFormatter::GREGORIAN,
+            'EEEE'
+        );
+
+        $this->fmtTime = new IntlDateFormatter(
+            $langCode,
+            IntlDateFormatter::FULL,
+            IntlDateFormatter::FULL,
+            date_default_timezone_get(),
+            IntlDateFormatter::GREGORIAN,
+            'HH:mm'
+        );
+
+        $this->trans = new Translator();
+        $this->trans->setLanguage($langCode);
+    }
+
     public function format(Offer $offer): string
     {
         $dateFrom = $offer->getStartDate()->setTimezone(new \DateTimeZone(date_default_timezone_get()));
